+++ 2lab/POS/src/main/java/com/team1206/pos/service/service/ServiceRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface ServiceRepository extends JpaRepository<Service, UUID> {
@@ -13,5 +14,5 @@ public interface ServiceRepository extends JpaRepository<Service, UUID> {
             "WHERE (:name IS NULL OR s.name LIKE %:name%) " +
             "AND (:price IS NULL OR s.price = :price) " +
             "AND (:duration IS NULL OR s.duration = :duration)")
-    Page<Service> findAllWithFilters(String name, Double price, Long duration, Pageable pageable);
+    Page<Service> findAllWithFilters(String name, BigDecimal price, Long duration, Pageable pageable);
 }
